backend/requirements.txt
Django==4.2.28
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==24.3
numpy==1.26.0
scipy==1.11.4
pandas==2.0.3

# Time series forecasting
statsmodels==0.14.1
scikit-learn==1.3.2

# Password hashing
argon2-cffi==23.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
# Additional dependencies for new features (all versions patched for security)
Pillow==10.3.0
statsmodels==0.14.1
# pandas==2.1.4

# MySQL database driver
pymysql==1.1.0


